Give query objects a home in Finders::
Query objects were split across two directories with no rule: two lived in
app/presenters/ next to genuine view-model code, two in app/models/ next to
ActiveRecord classes. "Where does a read-side query go" had two existing
answers, so it had none.

Finders:: is now the read-side counterpart to Ops::, wired the same way, so
the directory and the constant agree and the name says what the class is for.

Only pure query objects moved. The *Options presenters query too but exist to
build select options; CachedDashboard and ServerDashboard assemble a page's
view model; EffectiveConfig resolves inherited values. Those are not finders
and stayed put.

Finders::TwilightStruggle shadows ::TwilightStruggle, so the two moved
integration finders reference their models with a leading :: — the same trap
Ops::ServerConfiguration has.
