+++ config/application.rb
@@ -23,6 +23,9 @@ end
 module Bot
 end
 
+module Finders
+end
+
 module App
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
@@ -37,6 +40,8 @@ module App
     Rails.autoloaders.main.collapse(Rails.root.join("app/plugins/*/events"))
 
     Rails.autoloaders.main.push_dir(Rails.root.join("app/operations").to_s, namespace: Ops)
+
+    Rails.autoloaders.main.push_dir(Rails.root.join("app/finders").to_s, namespace: Finders)
 
     Rails.autoloaders.main.push_dir(Rails.root.join("app/bot").to_s, namespace: Bot)
     Rails.autoloaders.main.collapse(Rails.root.join("app/bot/events"))
