+++ docs/README.md
@@ -9,6 +9,7 @@ work actually touches.
 - [architecture/overview.md](architecture/overview.md) — the four processes (web, bot, jobs, postgres/redis), why the gateway bot can't boot inside Puma, AR pool sizing for unbounded handler threads, the web-writes/bot-reads config split, the three Discord credentials, and the `server`-not-`guild` naming rule.
 - [architecture/persistence.md](architecture/persistence.md) — prefixed-UUID string PKs generated by a Postgres column default, why snowflakes stay `bigint`, the `string_enum` macro, and the validation/constraint mirroring `active_record_doctor` enforces.
 - [architecture/operations-layer.md](architecture/operations-layer.md) — the `Ops::` write seam shared by bot and web: how operations are namespaced, why they take full objects instead of record IDs, and the `Result` contract.
+- [architecture/finders.md](architecture/finders.md) — the `Finders::` read seam: what counts as a query object and what stays a presenter, why the namespace mirrors `Ops::`, and the constant-shadowing trap that comes with it.
 - [architecture/config-propagation.md](architecture/config-propagation.md) — how a config change saved in the web UI reaches the running bot process over Redis pub/sub, the five event types on the bus, the `:ready` reconcile sweep that covers downtime, and how `BotSetting` differs from `ServerConfiguration`.
 
 ## Bot
