+++ docs/bot/notifications.md
@@ -12,7 +12,7 @@ The web side reads them cross-server, scoped to the session's authorized server 
 — notifications span all manageable servers.
 
 **Auth scope:** `NotificationsController` and `Notifications::ReadsController`
-include `SetsVisibleServers`. The query object `AuthorizedNotifications`
+include `SetsVisibleServers`. The query object `Finders::AuthorizedNotifications`
 (app/presenters/) joins `server_configurations` and filters by
 `managed_server_ids` from the session — the narrower of the two sets, so a
 tournament organiser who can open a server's dashboard still doesn't read its
